pyproject.toml
[tool.poetry]
name = "headhunter-best-vacancies"
version = "0.1.0"
description = ""
authors = ["KateRomanova <kate.romanova@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
